server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["saxenabhishek <as7122000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
python-dotenv = "^0.18.0"
autopep8 = "^1.5.7"
djongo = "^1.3.6"
schema = "^0.7.4"
dnspython = "^2.1.0"
gunicorn = "^20.1.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
